+++ noodles-tabix/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 0.12.0 - 2022-08-16
 
 ### Changed